feat(story-5.2): complete topic selection & speaker brainstorming implementation
Backend (GREEN phase):
- Implement topic selection with workflow integration (AC14-16)
- Add POST /api/v1/events/{eventCode}/topics endpoint
- Implement speaker pool management (AC9-13)
- Add POST /api/v1/events/{eventCode}/speakers/pool endpoint
- Add V15 migration for topic_id column in events table
- All 16 integration tests passing (100%)

Frontend fixes:
- Add eventType and workflowState to Event schema in OpenAPI spec
- Regenerate TypeScript types from updated OpenAPI spec
- Fix EventDto → Event type reference in workflowService
- Add eventType field to event import transformation
- Remove conflicting type definitions from EventUI/EventDetailUI
- Fix TFunction type parameter in EventCard

OpenAPI updates:
- Add eventType (FULL_DAY/AFTERNOON/EVENING) to Event schema
- Add workflowState (16-step workflow) to Event schema
- Add AddSpeakerToPoolRequest schema
- Add SpeakerPoolResponse schema

Code quality:
- Fix all checkstyle violations (line length, star imports, operator wrap)
- Fix ESLint warnings (TFunction type instead of any)

All TypeScript errors resolved. Build successful.

+++ services/event-management-service/src/main/resources/db/migration/V14__add_topic_speaker_pool_tables.sql
@@ -20,7 +20,7 @@ CREATE TABLE topics (
     usage_count INTEGER DEFAULT 0,
     staleness_score INTEGER DEFAULT 100 CHECK (staleness_score >= 0 AND staleness_score <= 100),
     calculated_wait_period INTEGER, -- in months
-    partner_influence_score NUMERIC(5,2),
+    partner_influence_score DOUBLE PRECISION,
     similarity_scores JSONB DEFAULT '[]', -- Array of {topicId: UUID, score: number}
     is_active BOOLEAN DEFAULT TRUE,
     title_vector tsvector GENERATED ALWAYS AS (to_tsvector('english', title)) STORED,
@@ -36,8 +36,8 @@ CREATE TABLE topic_usage_history (
     event_id UUID NOT NULL REFERENCES events(id) ON DELETE CASCADE,
     used_date TIMESTAMP WITH TIME ZONE NOT NULL,
     attendee_count INTEGER,
-    feedback_score NUMERIC(3,2),
-    engagement_score NUMERIC(3,2), -- 0.00 to 1.00
+    feedback_score DOUBLE PRECISION,
+    engagement_score DOUBLE PRECISION, -- 0.00 to 1.00
     created_at TIMESTAMP WITH TIME ZONE DEFAULT NOW(),
     updated_at TIMESTAMP WITH TIME ZONE DEFAULT NOW()
 );
